Allow direct transitions between stickyTop and stickyBottom modes
This commit adds checks to `handleScroll()` to allow for direct
transitions between the `stickyTop` and `stickyBottom` modes.

This fixes the problem reported in issue #24, where a programmatic
scroll (`window.scrollTo()`) issues a single event, that event
causes a state change (say from `stickyTop` to `relative`),
but does not actually reposition the `StickyBox` where it should be.